In mt32emu_qt: - Added support for using different sample rates in audio drivers using sample rate conversion.   It is still preferable to use generic sample rate of MT-32 compatible units 32000 Hz and sample rate conversion   should be used if a different sample rate is necessary (e.g. audio system doesn't support 32000 Hz) or deliberate.   For best performance, libsoxr is supported. Libsamplerate is also supported as it is widely available.   As a worst case, the integrated linear interpolator can be used (it's fast but the quality...).

+++ mt32emu_qt/src/audiodrv/AudioFileWriterDriver.cpp
@@ -40,9 +40,9 @@ void AudioFileWriterStream::close() {
 	writer.stop();
 }
 
-quint32 AudioFileWriterStream::estimateMIDITimestamp(const MasterClockNanos refNanos) {
+quint64 AudioFileWriterStream::estimateMIDITimestamp(const MasterClockNanos refNanos) {
 	MasterClockNanos midiNanos = (refNanos == 0) ? MasterClock::getClockNanos() : refNanos;
-	return quint32(((midiNanos - timeInfo[0].lastPlayedNanos) * sampleRate) / MasterClock::NANOS_PER_SECOND) + midiLatencyFrames;
+	return quint64(((midiNanos - timeInfo[0].lastPlayedNanos) * sampleRate) / MasterClock::NANOS_PER_SECOND) + midiLatencyFrames;
 }
 
 AudioFileWriterDevice::AudioFileWriterDevice(AudioFileWriterDriver &driver, QString useDeviceName) :
